Fixes after PR
