+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 1.5.6
 
 - Add markdown support. Thanks @felixgabler for the PR!
+- Add `previewTapOptions` that allow to configure to open link preview when tapped on preview's image or title. Thanks @felixgabler for the PR!
+- Add `dateHeaderBuilder`. Thanks @arsamme for the PR!
 - Add `onMessageVisibilityChanged` handler. Thanks @felixgabler for the PR!
-- Add `receivedMessageBodyLinkTextStyle` and `sentMessageBodyLinkTextStyle` to the theme. Thanks @felixgabler for the PR!
+- Add `receivedMessageBodyBoldTextStyle`, `receivedMessageBodyCodeTextStyle`, `receivedMessageBodyLinkTextStyle` and `sentMessageBodyBoldTextStyle`, `sentMessageBodyCodeTextStyle`, `sentMessageBodyLinkTextStyle` to the theme. Thanks @felixgabler for the PR!
 - Add `sendButtonMargin` to the theme. Thanks @damian-kaczmarek for the PR!
 - Add German localization. Thanks @felixgabler for the PR!
 - Update to Flutter 2.10.4. Requires Dart >= 2.16.0.
