+++ lib/src/widgets/attachment_button.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'inherited_chat_theme.dart';
 import 'inherited_l10n.dart';
