+++ lib/src/widgets/file_message.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
+
 import '../util.dart';
 import 'inherited_chat_theme.dart';
 import 'inherited_l10n.dart';
